Update packaging strategy
