+++ src/main/java/entity/inventory/DairySoy.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.inventory;
 
 /**
  * Entity class for dairy soy.
