+++ src/main/java/entity/inventory/HasFreshness.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.inventory;
 
 
 /**
