+++ src/main/java/entity/inventory/Inventory.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.inventory;
 
 /**
  * Interface inventory.
