+++ src/main/java/entity/inventory/SoftDrink.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.inventory;
 /**
  * Entity class for soft drinks.
  */
