+++ src/main/java/entity/kitchen/KitchenStaff.java
@@ -1,11 +1,13 @@
-package entity;
+package entity.kitchen;
+
+import entity.delivery.Staff;
 
 /**
  * Entity class for kitchen.
  */
 
 
-public class KitchenStaff extends Staff{
+public class KitchenStaff extends Staff {
     public KitchenStaff(String staffID, String staffName, String password, int staffSalary) {
         super(staffID, staffName, password, staffSalary);
     }
