+++ src/main/java/entity/orderList/Order.java
@@ -8,7 +8,6 @@ package entity.orderList;
  */
 
 import constant.orderSystem.ItemStatus;
-import entity.Dish;
 
 import java.util.ArrayList;
 import java.util.HashMap;
