+++ src/main/java/use_case/customerSystem/PlaceOrder.java
@@ -9,7 +9,7 @@ package use_case.customerSystem;
 import use_case.dishList.DishList;
 import use_case.kitchen.OrderQueue;
 import use_case.boundary.PlaceOrderInputBoundary;
-import entity.Dish;
+import entity.orderList.Dish;
 import entity.orderList.Order;
 
 import java.util.ArrayList;
