+++ src/main/java/use_case/reviewList/ReviewList.java
@@ -1,12 +1,9 @@
 package use_case.reviewList;
-import entity.*;
-import gateway.ReviewReadWriter;
+import entity.review.Review;
 
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 
 public class ReviewList implements Serializable {
     private static HashMap<Integer, ArrayList<Review>> reviews;
