+++ src/test/java/UserListTest.java
@@ -5,10 +5,12 @@
  */
 
 import constant.mangerSystem.UserType;
-import entity.*;
-import entity.DeliveryStaff;
-import entity.ServingStaff;
-import entity.InventoryStaff;
+import entity.delivery.DeliveryStaff;
+import entity.delivery.ServingStaff;
+import entity.customer.Customer;
+import entity.inventory.InventoryStaff;
+import entity.kitchen.KitchenStaff;
+import entity.manager.Manager;
 import org.junit.Before;
 import org.junit.Test;
 import use_case.userList.UserList;
