+++ src/test/java/controller/StaffControllerTest.java
@@ -2,9 +2,9 @@ package controller;
 
 import controller.staffSystem.StaffController;
 import entity.*;
-import entity.DeliveryStaff;
-import entity.ServingStaff;
-import entity.Dish;
+import entity.delivery.DeliveryStaff;
+import entity.delivery.ServingStaff;
+import entity.orderList.Dish;
 import entity.orderList.Order;
 import org.junit.Before;
 import org.junit.Test;
