+++ src/test/java/entity/OrderTest.java
@@ -1,6 +1,7 @@
 package entity;
 
 import constant.orderSystem.ItemStatus;
+import entity.orderList.Dish;
 import entity.orderList.Order;
 import org.junit.Before;
 import org.junit.Test;
